renamed resultList and resultListItem and fixed all the errors
